requirements.txt
coverage==3.7.1
coveralls==0.5
mockito==0.5.2
nose==1.3.4
pyserial==2.7
